requirements.txt
Flask==2.3.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.2
werkzeug>=2.0,<3.0